refactor: split matcher to multiple files, create appropriate directories
