+++ src/rxjs/hot-observable.ts
